RssAgent: Migrate from FeedNormalizer to Feedjira
FeedNormalizer is no longer maintained, and its Atom support has flaws
in that it throws away what RSS::Parser returns and falls back to using
SimpleRSS which is not capable of handling XML entities, resulting in
getting ususable URLs such as ones including `&amp;`.

A new key `links` is added, which lists all `link` elements.

A breaking change is removal of the `clean` option which needs to be
reimplemented later if it is important.

This should address #889 and #955, among others.